Optimise logic
